Removed TabButton.editable setting tabButtonCell.isEditable, as it caused a graphical glitch when right clicking editable tabs.
Overriding the TabButton's rightMouseDown to not call super worked, but that's not the root of the issue.  I only have time to dig as far as the tabButtonCell's isEditable parameter.  Setting this to false (or removing the ability to set it as I have) fixes the issue but still preserves editing.